dequeue: always clean up
Currently, the CleanUp method is only executed if an object was
_neither_ synced _nor_ marked. This is problematic, because it means
that objects that are already correctly stored will not be cleaned up,
even though they should not be re-processed. This commit fixes the logic
for the cleanup while also simplifying the function to make the logic
easier to follow, e.g. now, the code only ever has to call `mark` in one
place.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>