storage: move stat to storage
This commit adds a Stat method to the storage.Storage interface. This
allows the dequeuer to check if files need to be uploaded before doing
so. Currently the s3 storage performs this check internally, avoiding
extra uploads, however moving this to the dequeuer would allow all
storages to benefit from this functionality without needing to implement
the check themselves.

Unfortunately for the s3 storage, using done files complicates this
flow. There are two options:
1. Perform all of the stats again inside of the Store method to check
   what needs to be created. This means that in the worst case, a
   single upload can result in 4 stat calls, instead of the previous 2;
   or
2. Allow creating meta objects in the Stat method if the object exists
   but the meta object does not.

This commit takes the second approach, which simplifying the Store
method significantly, although it means that the Stat method has side
effects.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>