improve metrics: standardize names, add counters
This commit standardizes the way that metrics are named and labeled.
Now, this library uses the common pattern of having a single counter
with either `result="success"` or `result="error"` labels, rather than
two different counters.

Furthermore, this commit adds a new time series family to track storage
operations. This is used by wrapping and `Storage` implementation with
the function `NewInstrumentedStorage`.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>