introduce the ingest command
This commit modifies ingest to be a project in its own right with a
binary that manages ingestion of objects from source APIs to destination
APIs that is completely configurable with a single file. The motivation
is that this should simplify deployment of ingestion services.
Previously, creating a new ingestion service required building and
maintaining a separate repository, binary, and container for every
permutation of APIs, e.g. s3-drive, drive-s3, etc. Now, a single plugin
can be defined for any API and APIs can be mixed and matched as desired
and specified via configuration of the ingest command.
This will greatly simplify deployments, making it easier to maintain
and read the Kubernetes configuration, as previously having multiple
s3-s3 jobs would result in many similar deployments, whose inheritance
could get tricky to handle.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>