storage: create multi-storage
This commit adds a new multi-storage implementation of the Storage
interface. This storage system will store data on all underlying
storages, allowing a single dequeuer to store data in multiple
destinations. This is necessary so that a single message, published on
the subject for one single workflow, can be used to create objects in
multiple backends.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>