storage: allow client to store in shared drives
Currently, the Google Drive client is only configured to be able to
store objects in non-shared drives, even though it is configured to find
objects in shared drives. This commit aligns this behavior. Eventually
we may want to make this a configuration option for the storage that is
configurable though the destination yaml.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>