go-plugin (#47)
* Remove identifiable interface

The interface did not help and using a concrete type enables us to use
go-plugin, because it is harder to serve interfaces over rpc than
concrete types.

* Remove Object interface

Interfaces are more difficult to use as arguments in rpc.

* go-plugin: use hashicorp's go-plugin

This commit removes the usage of the stdplugin system. The tight
constrains about library versions blocked plugins from being loaded
because indirect dependencies did not match.

Hashicorp's go-plugin uses rpc and the constrains to load plugins are
less tight.

* README: correct comments

Remove references to the T object, remove Identifiable from the README
and regenerate the README.

* Use golangci-lint

I failed to find a way to ignore lint errors from govet, so I replaced
all linters with golangci-lint. It includes govet and staticcheck but
support inline comments to ignore some errors like `lostcontext`.

* Enable unit tests on other OS and arch

* Refactor of plugin package

Move all rpc related code to another file and remove passing errors
messages via rpc arguments but instead return an error on the rpc server
methods.

* Terminate plugins in unit tests

Especially in the config tests, plugins were not termintated. Maybe
because the after the context was canceled the main process exited
without giving the go routine enough time to kill the rpc server.

* Apply suggestion from code review

 - add mocks/nexter to gen-mock phony target
 - rename SimpleCodec to Codec
 - fix some comments
 - unexport some values not needed by importing packages

* storage interface

Keep the original storage interface and implement plugin related logic
in the plugin's main packages.

* plugin/noop.go: rename Destination implementation

Rename `NoopStorage` to `NoopDestination` because is implements the
Destination interface, not just the Storage interface.

* Timeout in plugin calls

We can't pass a context over RPC to the plugin server. So we set a fixed
timeout during configuration. A `context.NewWithTimeOut(timeout)` will
be passed to the implementations of the Source and Destination interfaces.

* Remove timeout

Tests were failing because the Next method needed the context of Reset
to not be canceled before Next was called until the next Reset.
This is due to the implementation of the s3 plugin.

* Apply suggestions from code review

Co-authored-by: Lucas Servén Marín <lserven@gmail.com>

* Copy only once in mutistorage

Copy the data from the source to a buffer that can be used by multiple
readers (Destinations) at the same time.

Add a test with a 100MB file that has to be written to two destinations.

* Update ingest.go

* Update README.md

Co-authored-by: Lucas Servén Marín <lserven@gmail.com>