PluginManager: start and stop plugins
Use a manager to start and stop plugins. The manager keeps track of its
plugins and can `Watch()` them and returning errors when a plugins is
not available anymore. Furthermore it can `Stop` all plugins. This way
we don't have zombie processes from the rpc servers that were not
terminated after exiting the main function.

Signed-off-by: leonnicolas <leonloechner@gmx.de>