observability: label dequeuer metrics with source
Right now, the metrics concerning the dequeuer include labels for the
workflow and the destination but not the source. We can naturally infer
the source from the workflow name, but it's nice to have this in one
place in the metrics to help us easily tell what documents may be
failing to copy without needing to refer to the configuration.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>