manage plugins (#52)
* Individually start Source and Destination

Before we had one function to intitialize the Source and Destination for
one plugin. This always created two processes. With the new way, we only
create one process per Source/Destination.

Signed-off-by: leonnicolas <leonloechner@gmx.de>

* PluginManager: start and stop plugins

Use a manager to start and stop plugins. The manager keeps track of its
plugins and can `Watch()` them and returning errors when a plugins is
not available anymore. Furthermore it can `Stop` all plugins. This way
we don't have zombie processes from the rpc servers that were not
terminated after exiting the main function.

Signed-off-by: leonnicolas <leonloechner@gmx.de>

* Clean up

 - comment on PluginManager.Stop
 - lock mutex in Stop and don't ever unlock because the Manager is not
   usable anymore.
 - logging in main method to start PluginManager.Watch()
 - cancel enqueuer context in run.Group (This was a work around to allow
   it more time to kill all rpc clients)

Signed-off-by: leonnicolas <leonloechner@gmx.de>
Co-authored-by: Lucas Servén Marín <lserven@gmail.com>