storage: use multi-storage
Currently, the multi-storage implementation is unused, which means that
workflows with multiple destinations _cannot_ enable cleanup or else
files may be deleted in the source before every storage has had a chance
to upload it.

To work around this limitation, let's re-enable the use of
storage.MultiStorage. Note: although the consumer the name and number of
consumers has changed, elements do not need to be re-queued. This
convenience is enabled by the fact that each consumer has its own view
of the stream for the workflow. Old consumers suffixed with
`__<destination>` will need to be manually deleted from NATS.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>