Makefile: ensure consistent go version
Currently, the generation of Go files produces a diff in CI because the
version of Go used to compile Mockery is different, resulting in some
different behavior during mock generation time. This commit modifies the
Makefile to ensure that all `go build` commands use a consistent version
of Go specified in the Makefile as a Docker image.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>