config: allow skipping failed workflows (#63)
Sometimes one workflow will be misconfigured, or have some runtime
issue, e.g. when Google Drive permissions are broken, which prevents the
entire ingest binary from starting. This is not always desirable, as
sometimes it is better for some ingestion to work rather than no
ingestion. This commit adds a flag to the binary that allows
misconfigured workflows to be skipped rather than causing the
application to exit with a failure. Whenever workflows are skipped, a
newly added metric is incremented, allowing administrators to alert on
ingest configuration errors rather than blindly running ingest without
realizing that a workflow is not running. This new mode of skipping
failed workflows is disabled by default.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>