prefer passing gatherer (#69)
* plugins: prefer passing Gatherer over Collectors

This commit changes the call signature of the plugin servers from
accepting a variadic of Prometheus Collectors to accepting a Gatherer.
This is more useful because it means that the caller can pass a
Prometheus Registry, which can also be passed by the caller to other
code that wants to register metric collectors. This is more convenient
than expecting Collectors, because metrics are normally private
implementation details of sub-packages and it would not be reasonable to
expect the main func of the plugin to be able to access or even know
about all descendent metrics. It inverts the way that metrics have to be
registered. Now, whenever a descendent package wants to register a new
metrics, they can simply register it, without having to coordinate with
the top level of the plugin's main func.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* drive storage: register drive operations metric

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

---------

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>