multi-storage: fail fast in during stat (#70)
Right now, if one of the storages backing a multi-storage returns an
error, the multi-storage still waits for all other storages to finish
responding before returning an error. This is wasteful, since we already
know that the final response of the stat call will be an error. This
commit changes this behavior so that all in-flight stats are canceled
as soon as one error is received.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>