plugins/drive: fix panic on startup
Currently, the Google Drive panic panics on startup because the find
method which requires a counter to be incremented, is called before the
counter is initialized.

This commit fixes this issue as well as fixes an issue where the
registry was never passed to the plugin, meaning the and metrics would
never actually be registered.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>