Use default Gatherer
Define a default Gatherer in cmd.go, where we also set the default
logger. This way we set those default values in the same file and same
function. Also it gives plugin developers more control over the
Gatherer. For example they can override it to avoid using the go
collector and process collector.

Signed-off-by: leonnicolas <leonloechner@gmx.de>