plugin: use options to set logger and gatherer (#78)
* plugin: use opstions to set logger

When starting the plugin server, allow the developer to set the
Registerer and the logger with an Option. This way plugins can use that
defined logger in there own manner. Before it was not possible to write
custom log messages.

Signed-off-by: leonnicolas <leonloechner@gmx.de>

* Use default Gatherer

Define a default Gatherer in cmd.go, where we also set the default
logger. This way we set those default values in the same file and same
function. Also it gives plugin developers more control over the
Gatherer. For example they can override it to avoid using the go
collector and process collector.

Signed-off-by: leonnicolas <leonloechner@gmx.de>

---------

Signed-off-by: leonnicolas <leonloechner@gmx.de>