plugin/manager.go: skip on failed Gather
If a plugin does not return valid prometheus metrics for the Gather
call, we were failing the whole operation for all plugins. So one
failing call makes it impossible to collect any metrics from ingest.
This commit will log and skip that plugin.
To follow up, we should collect metrics of the rpc client to see if
Gather calls are failing.

Signed-off-by: leonnicolas <leonloechner@gmx.de>